_sbrk: use main sp instead of current sp to check heap & stack collision
_sbrk is used mainly by malloc() to allocate heap area. Before
allocating more space to heap or memory from heap, it is normal practice
to make sure this allocation does not overlap with stack pointer (or SP
+ min_space_for_stack).

This check passes when your float printf() is in main.c  as the stack
pointer in this case of MSP which is at the top of RAM (- stack in use).
The same check when done from task/thread fails as the stack pointer
(PSP) is assigned space in the heap if dynamic allocation is
chosen(heap_x.c) or in text area when static allocation is chosen
(configSUPPORT_STATIC_ALLOCATION). In both cases, PSP is below heap area
& is treated as heap/stack collision.

The correct implementation is to check heap/stack collision/overlap with
MSP.

Signed-off-by: Vikas Manocha <vikas.manocha@st.com>